refactor badge files
