+++ blotztask-mobile/src/i18n/locales/en/badge.json
@@ -1,0 +1,4 @@
+{
+  "achievementUnlocked": "Achievement unlocked!",
+  "obtainedOn": "Obtained on {{date}}"
+}
