+++ blotztask-mobile/src/i18n/locales/zh/badge.json
@@ -1,0 +1,4 @@
+{
+  "achievementUnlocked": "成就解锁！",
+  "obtainedOn": "获得于 {{date}}"
+}
