+++ blotztask-mobile/src/i18n/locales/zh/common.json
@@ -47,9 +47,5 @@
     "title": "哎呀，出问题了！",
     "message": "出了点小问题，重启App试试看！",
     "button": "重启应用"
-  },
-  "badge": {
-    "achievementUnlocked": "成就解锁！",
-    "obtainedOn": "获得于 {{date}}"
   }
 }
